Review changes

+++ modules/libScePad/cconfig.h
@@ -7,7 +7,7 @@ class ControllerConfig {
   struct Key {
     uint16_t key;
     uint16_t mod;
-  } m_keymap[ControllerKey::CONTROLLER_KEYS_COUNT];
+  } m_keymap[(uint32_t)ControllerKey::CONTROLLER_KEYS_COUNT];
 
   struct DeadZone {
     float x, y;
@@ -21,7 +21,7 @@ class ControllerConfig {
   public:
   auto GetPadType(int n) const { return m_pads[n].type; }
 
-  auto GetBind(ControllerKey key) const { return m_keymap[key]; }
+  auto GetBind(ControllerKey key) const { return m_keymap[(uint32_t)key]; }
 
   ControllerConfig();
   virtual ~ControllerConfig() = default;
